add scroll to active row
